+++ tests/__main__.py
@@ -6,6 +6,7 @@ from tests.test_complex_operations import *
 from tests.test_deserialization_from_formal import *
 from tests.test_deserialization_from_sympy import *
 from tests.test_add_simplification import *
+from tests.test_mul_simplification import *
 
 
 if __name__ == "__main__":